Moved styles into the new directory
